+++ src/sources/utils.ts
@@ -96,68 +96,3 @@ export function isTextAvatar(sourceType:string):boolean{
     return ["INITIALS","VALUE"].indexOf(sourceType) > -1;
 
 }
-
-/**
- * Extract avatar image from the given data object.
- * The extraction path is based on the sourceType.
- * 
- * @export
- * @param {string} sourceType 
- * @param {*} data 
- */
-export function extractAsyncAvatarData(sourceType:string,data:any,size:number){
-    switch (sourceType) {
-        case "GOOGLE":
-            return _extractGoogleAvatar(data,size);
-        case "VKONTAKTE":
-            return _extractVkontakteAvatar(data);
-        case "GITHUB":
-            return _extractGithubAvatar(data, size);
-        default:
-            // source not defined
-            return undefined;
-    }
-}
-
-/**
- * Extract google avatar from json data
- * 
- * @param {*} data 
- * @returns 
- * @memberof AvatarComponent
- */
-function _extractGoogleAvatar(data: any,size:number) {
-    const avatarSrc = data.entry.gphoto$thumbnail.$t;
-    if (avatarSrc) {
-        return avatarSrc.replace('s64', 's' +size);;
-    }
-}
-
-/**
- * extract vkontakte avatar from json data
- * 
- * @param {*} data 
- * @returns 
- * @memberof AvatarComponent
- */
-function _extractVkontakteAvatar(data: any) {
-    // avatar key property is the size used to generate avatar url
-    // size property is always the last key in the response object
-    const sizeProperty = Object.keys(data["response"][0]).pop();
-    // return avatar src
-    return data["response"][0][sizeProperty];
-}
-
-/**
-* extract github avatar from json data
-*
-* @param {*} data
-* @returns url of the github avatar icon
-* @memberof AvatarComponent
-*/
-function _extractGithubAvatar(data: any, size?: number) {
-    if (size) {
-        return `${data.avatar_url}&s=${size}`;
-    }
-    return data.avatar_url;
-}
